Evaluation

Examples

Structs

squirrel.seval.TestCases — Type

The TestCases datatype may be populated by the associated function of the form:

TestCases( par::NTuple , N::Int , np::Int , X::Array{RealMtx,1} , Xtar::Array{RealVec,1} )

The variables are defined in the following way:

par     ::  NTuple              # Parameter tuple
N       ::  Int                 # Number of generated test cases
np      ::  Int                 # Number of emission points
X       ::  Array{RealMtx,1}    # Emission points
Xtar    ::  Array{RealVec,1}    # Target point

squirrel.seval.TestData — Type

The TestData datatype may be populated by the associated function of the form:

TestData(   par::NTuple , N::Int    , 
            X::Array{RealMtx,1}     ,   Xtar::Array{RealVec,1}     ,
            Xc::Array{RealVec,1}    ,   Xsc::Array{RealVec,1}      ,
            erh::RealVec    ,   erv::RealVec    ,   err::RealVec   ,
            erhC::RealVec   ,   ervC::RealVec   ,   errC::RealVec  ,
            Xc2::Array{RealVec,1}   ,   Xsc2::Array{RealVec,1}     ,
            erh2::RealVec   ,   erv2::RealVec   ,   err2::RealVec  ,
            erhC2::RealVec  ,   ervC2::RealVec  ,   errC2::RealVec )

The quantities are defined in the following way:

par     ::  NTuple              # Parameter tuple
N       ::  Int                 # Number of cases
X       ::  Array{RealMtx,1}    # Emission points
Xtar    ::  Array{RealVec,1}    # Target point
Xc      ::  Array{RealVec,1}    # Flat spacetime position
Xsc     ::  Array{RealVec,1}    # Squirrel position
erh     ::  RealVec             # Horizontal error relative to Xsc
erv     ::  RealVec             # Vertical error relative to Xsc
err     ::  RealVec             # Total error relative to Xsc
erhC    ::  RealVec             # Horizontal error relative to Xc
ervC    ::  RealVec             # Vertical error relative to Xc
errC    ::  RealVec             # Total error relative to Xc
Xc2     ::  Array{RealVec,1}    # Flat spacetime position     (Aux.)
Xsc2    ::  Array{RealVec,1}    # Squirrel position      (Auxiliary)
erh2    ::  RealVec             # Horizontal errors      (Auxiliary)
erv2    ::  RealVec             # Vertical error         (Auxiliary)
err2    ::  RealVec             # Total error            (Auxiliary)
erhC2   ::  RealVec             # Horizontal errors      (Auxiliary)
ervC2   ::  RealVec             # Vertical error         (Auxiliary)
errC2   ::  RealVec             # Total error            (Auxiliary)

The quantities suffixed with 2 are auxiliary quantities which are needed in the four-point case, in which the special relativistic location algorithms generally suffer from the bifurcation problem.

Test case generation

Point generation

squirrel.seval.pgen — Function

pgen( Rs::Real , gfunc::Function , tol::Real , np::Int=4 , 
      Δψ::Real=Δψ0 )

The pgen function generates a target point Xtar and np emission points in a 4×np matrix X such that the points in X lie on the past light cone of Xtar with respect to the metric gfunc, and have spatial radii values of ~Rs (defined with respect to spatial origin). The parameter tol is the tolerance parameter for the integration. This function returns the tuple (X,Xtar).

Multiple case generator

squirrel.seval.gen — Function

gen( N::Int , g::Function , np::Int=4 , Δψ::Real=Δψ0 
          , tpfl::DataType=Float64 
          , REs ::  Real=1.4365276211950395e9       # Earth radius
          , RR  ::  Real=1.4365277e9    # Just above Earth surface
          , Rs  ::  Real=6e9            # Satellite radius
          , tolh::  Real=1e-14          # Tolerance for ODE solvers
          , tol ::  Real=1e-10          # Tolerance for ODE solvers
        )

The gen function generates N sets of target points Xtar and emission points X with respect to the metric gfunc. This function returns a quantity of the TestCases datatype.

Evaluation

squirrel.seval.main — Function

main(  tc::TestCases , sloc::Function , g::Function , Nx::Int=-1 , 
       tpfl::DataType=Double64 , tol::Real=1e-10 , ξ::Real=2e1 , 
       nb::Int=24 , ne::Real=6 )

This is the main evaluation function. It takes as input a collection of test cases contained in a quantity of datatype TestCases, and evaluates the squirrel locator function sloc on each test case relative to a given metric function g. The function main returns results in a quantity of type TestData.

The variable Nx is a limiter for the number of test cases to run. The variable tpfl specifies the floating point precision for the calculations. variables tol, ξ, nb, and ne are inputs for the sloc function, and correspond respectively to the integration tolerance, outlier detection threshold, number of steps for the Broyden algorithm, and number of emission points.