Geodesic solver

Null Geodesics

Null geodesics in a spacetime geometry described by a metric $g_{μν}$ are described by the Hamiltonian:

\[H = \frac{1}{2} g^{μν} p_μ p_ν\]

and the associated Hamilton equations:

\[\frac{dx^μ}{dλ} = \frac{∂H}{∂p_μ}\]

\[\frac{dp_μ}{dλ} = - \frac{∂H}{∂x^μ}.\]

The conjugate momenta $p_μ$ are defined as:

\[p_μ := g_{μν} \frac{dx^ν}{dλ},\]

and for null geodesics, the initial data satisfies:

\[\left.δ{^i}{_μ} \frac{dx^μ}{dλ}\right|_{λ=0} = {\bf v}^i\]

\[\left. g_{μν}({x}_0) \frac{dx^μ}{dλ}\frac{dx^ν}{dλ} \right|_{λ=0} = 0.\]

The solution to Hamilton's equations is formally given by $x^μ=x^μ(λ,{x}_0,{\bf v})$, and since $λ$ (being an affine parameter) can be redefined linearly, it is appropriate to set up the problem so that $λ∈[0,1]$, with $λ=0$ being the initial point and $λ=1$ is the final point.

Implementation

Hamiltonian

The Hamiltonian function takes the form:

squirrel.HamGeo — Function

HamGeo( Z::RealVec , gfunc::Function )

The HamGeo function computes the geodesic Hamiltonian $H = rac{1}{2} g^{μν} p_μ p_ν$ from the metric $g_{μν}$ provided in the function gfunc and the spacetime position x and momentum p encoded in Z=(x,p).

Hamilton's equations

Hamilton's equations may be written in terms of the phase space coordinate $z^α$, where $z=(x,p)$.

\[\frac{dz^α}{dλ} = J^{αβ} \frac{∂H}{∂z^β}.\]

where $J^{αβ}$ is the symplectic matrix, which has the form:

\[J = \left[ \begin{array}{cccc} O & I \\ -I & O \end{array} \right].\]

Initial data

squirrel.nullenforcerp — Function

nullenforcerp( V0::RealVec , X::RealVec , gfunc::Function )

The nullenforcerp function takes a vector V0 and modifies the time or first component so that it is null at the point X with respect to gfunc and past directed; the function returns the modified vector.

Missing docstring.

Missing docstring for squirrel.nullenforcerp. Check Documenter's build log for details.

Geodesic functions

Missing docstring.

Missing docstring for squirrel.HamGeo. Check Documenter's build log for details.