Metrics

Kerr-Schild metric

The Kerr-Schild metric takes the form:

\[ g_{\mu \nu} = \eta_{\mu \nu} + f \, k_\mu \, k_\nu \]

where

\[ k_\mu = \left( 1 , \frac{r \, x + a \, y}{r^2+a^2} , \frac{r \, y - a \, x}{r^2+a^2} , \frac{z}{r}\right)\]

\[ f = \frac{2 \, G \, M \, r^3}{r^4+a^2 \, z^2},\]

and $r$ is implicitly defined by:

\[\frac{x^2+y^2}{r^2+a^2} + \frac{z^2}{r^2} = 1 .\]

The components of the Kerr-Schild metric can be calculated using gks function in general

squirrel.metric.gks — Function

gks( X::RealVec , a::Real=0 , GM::Real=1 )

The gks function takes a point X in Cartesian Kerr-Schild coordinates and calculates the components of the Kerr-Schild metric at that point in a Kerr spacetime with rotation parameter a and with the product of gravitational constant and mass GM

or using the ge function for Earth-like parameters

squirrel.metric.ge — Function

ge( X::RealVec )

The ge function takes a point X in Cartesian Kerr-Schild coordinates and calculates the components of the Kerr-Schild metric at that point in a Kerr spacetime with relation between mass and angular momentum roughly corresponding to the Earth value ($a=738GM$)

Weak field metric

For solar system and terrestrial positioning, the weak field metric suffices. The weak field metric has the form:

\[g_{\mu \nu} = \eta_{\mu \nu} - 2 \, V \, \delta_{\mu \nu},\]

where $V$ is the gravitational potential.

Gordon metric

To incorporate atmospheric and ionospheric effects, one uses the analogue Gordon metric, which takes the form (with $g_{\mu \nu}$ being the gravitational metric):

\[ \bar{g}_{\mu \nu} = g_{\mu \nu} + \left(1-\frac{1}{n^2}\right) u_\mu u_\nu ,\]

The general Gordon metric is implemented in the following function:

squirrel.metric.gGordon — Function

gGordon( X::RealVec , n::Function=nIR , gfunc::Function=giso , 
         U::RealVec=Float64[-1;0;0;0] )

The function gGordon computes the Gordon metric. The argument n can be replaced with a user-supplied index of refraction function, the argument gfunc can be replaced with a user-supplied background metric, and the argument U can be replaced with a user-supplied fluid four-velocity function.

In principle, the function implemented above can be used with any user defined index of refraction profile. However, for evaluation purposes, we implement some simplified index of refraction profiles in the following sections.

Index of refraction models

The index of refraction is written in the following form:

\[n = 1 + \Delta n_{\rm atm} + \Delta n_{\rm ion}\]

Atmospheric component

The atmospheric component for the index of refraction is given here as a function of the height $h$ from the surface of the Earth:

\[\Delta n_{atm} = \frac{A_1}{B_1 + C_1 \left(h-H_1)\right)} + \frac{A_2}{B_2 + C_2 \left(h-H_2)\right)}\]

with the parameter choices:

\[\begin{aligned} A_1 &= -222.666 &\qquad A_2 &= -253.499 \\ B_1 &= 99.0621 &\qquad B_2 &= 112.757 \\ C_1 &= 0.157823 ~\text{km}^{-1} &\qquad C_2 &= 0.179669 ~\text{km}^{-1}\\ H_1 &= -7.1541 ~\text{km} &\qquad H_2 &= -7.15654 ~\text{km} \end{aligned}\]

The function $\Delta n_{atm}$ is implemented in the following:

squirrel.metric.ΔnatmStd — Function

ΔnatmStd( h::Real )

The function ΔnatmStd is a profile for the atmospheric index of of refraction obtained from the U. S. standard atmosphere data. The argument h is height from the Earth's surface in km.

Ionospheric component

The ionospheric component index of refraction is given by the following:

\[\begin{aligned} \Delta n_{ion} \approx \left(4.024 \times 10^{-17}\right) [N_e/\text{m}^{-3}] \end{aligned}\]

where the electron density is given by:

\[\begin{aligned} N_e :=& \biggl[\alpha_D \, \tilde Ep(h,h_D,b_D) + \alpha_E \, \tilde Ep(h,h_E,b_E) + \alpha_F \, \tilde Ep(h,h_F,b_F)\biggr], \end{aligned}\]

with the parameter choices:

\[\begin{aligned} \alpha_D &= 10^{12}~\text{m}^{-3} &\quad h_D &= 75~\text{km} &\quad b_D &= 5~\text{km} \\ \alpha_E &= 2.5 \times 10^{11}~\text{m}^{-3} &\quad h_E &= 130~\text{km} &\quad b_E &= 30~\text{km} \\ \alpha_F &= 10^{11}~\text{m}^{-3} &\quad h_F &= 300~\text{km} &\quad b_F &= 50~\text{km} . \end{aligned}\]

The function $\tilde Ep(h,h_E,b_E)$ is the pseudo-Epstein function:

\[\begin{aligned} \tilde Ep(h,h_c,B) := &\frac{1}{16} \biggl\{ \left[1+\left(\tfrac{h-h_c}{2 B}\right)^2\right]^{-1} +\left[1+\left(\tfrac{h-h_c}{4 B}\right)^2\right]^{-2} +\left[1+\left(\tfrac{h-h_c}{6 B}\right)^2\right]^{-3} +\left[1+\left(\tfrac{h-h_c}{7 B}\right)^2\right]^{-4} \biggr\}^2, \end{aligned}\]

defined as:

squirrel.metric.pEp — Function

pEp( h , hc , B , Nmax )

The function pEp implements the pseudo-Epstein line shape function used to approximate the Epstein function used to construct the ionospheric electron density profiles.

The ionospheric index of refraction, also written in terms of the height $h$ from the surface of the Earth, is implemented in the following:

squirrel.metric.Δnios — Function

Δnios( h::Real , θ::Real , ϕ::Real )

The function Δnios provides a crude model for the ionospheric electron density profile. The variable h represents height from the Earth's surface in units of km.

The unperturbed total index of refraction $\Delta n_{\rm tot} = \Delta n_{\rm atm} + \Delta n_{\rm ion}$ is given in the following (in surface-adapted coordinates):

squirrel.metric.Δntot — Function

Δntot( h::Real , θ::Real , ϕ::Real )

The function Δntot is the total index of refraction profile, which is a sum of atmospheric and ionospheric index of refraction profiles. The variable h represents height from the Earth's surface in units of km.

Perturbation models

To model uncertainties in the index of refraction profile, perturbations of the following form are introduced (also in surface-adapted coordinates):

\[\begin{aligned} \Delta n_{\rm pert} &= \Delta n_{\rm atm} \left(1 + \delta_1 \, \tilde p_1(h) \right) \\ & \qquad + \Delta n_{\rm ion} \left(1 + \delta_2 \, \tilde p_2(h)\right), \end{aligned}\]

and implemented in the following.

squirrel.metric.Δntp — Function

Δntp( h::Real , θ::Real , ϕ::Real , δ1::Real=0.001
           , δ2::Real=0.010 , Patm::Function=h->1.0
           , Pion::Function=h->1.0 )

The function Δntp adds the perturbations provided in the perturbation model to the atmospheric and ionospheric profiles defined earlier.

The respective perturbation profile function $p_A$ and line shape function $Ls$ are defined:

\[\tilde p_A(h) = \sum_{i} \alpha_i \tilde Ls(h,h_{0,i},\sigma_i)\]

squirrel.metric.P — Function

P( h::Real , h0::RealVec , σ::RealVec )

The function P constructs the perturbation model, which consists of a sum of line shape functions LSF, with the parameters provided in the vector arguments h0 and σ.

\[Ls(h,h_{0},\sigma) = \frac{\sigma^2}{\sigma^2 + (h-h_0)^2} \frac{\sigma^4} {\sigma^4 + (h-h_0)^4}\]

squirrel.metric.LSF — Function

LSF( x , xo , σ )

The function LSF implements a line shape function employed in the perturbation models; it has a faster falloff than the Lorentzian.

The parameters $h_i$ and $\sigma_i$ are fed into the perturbation profile functions as vectors.

Conversion functions

Index of refraction functions are converted from surface-adapted coordinates to spherical coordinates with the following function:

squirrel.metric.nIRs — Function

nIRs( RE::Real , r::Real , θ::Real , ϕ::Real , Δnf::Function )

The function nIRs computes the index of refraction in spherical coordinates from the index of refraction profile function Δnf in spherical coordinates.

Finally, the index of refraction functions are converted to Cartesian coordinates with the following function:

squirrel.metric.nIR — Function

nIR( X::RealVec , Δnf::Function=Δntot )

The function nIR computes the index of refraction in Cartesian coordinates; the profile is anchored to the WGS84 ellipsoid defined by the rell function.