Squirrel algorithm

Root finding algorithm

Given some function $f(x)$, the standard approach is the Newton method, which finds the roots according to the prescription:

\[ x_{i+1} = x_i + {\bf J}^{-1}_{i} f(x_i)\]

where ${\bf J}^{-1}_{i}$ is the inverse of the Jacobian matrix ${\bf J}$ for $f(x)$ evaluated at $x_i$. However, in situations which the computation of the Jacobian matrix ${\bf J}$ becomes expensive, one may instead employ the Broyden method, which is a quasi-Newton root finding method. The Broyden algorithm involves first computing the Jacobian matrix ${\bf J}$ and its inverse for $f(x)$. However, instead of computing the Jacobian matrix at each iteration (as is done in the Newton method), the Broyden algorithm updates the (inverse) Jacobian matrix according to the Sherman-Morrison update formula:

\[ {\bf J}^{-1}_{i+1} = {\bf J}^{-1}_{i} + \frac{Δv^{T}_i - {\bf J}^{-1}_{i} Δf_i } {Δv^{T}_i {\bf J}^{-1}_{i} Δf_i } Δv^{T}_i {\bf J}^{-1}_{i},\]

which is implemented in the following function:

squirrel.JiSMU — Function

JiSMU( ΔF::RealVec , Δx::RealVec , Ji::RealMtx )

The function JiSMU implements the Sherman-Morrison update formula.

The Broyden update formula is implemented in the following function:

squirrel.bsolve — Function

bsolve( F::Function , J::RealMtx , f0::RealVec , x0::RealVec ,
        nb::Int=24 )

The function bsolve implements the Broyden algorithm; in particular, it finds the roots of the function F(x), given the Jacobian matrix J and the initial guesses f0 and x0. The parameter nb specifies the maximum number of iterations.

Geodesic integration