Geodesic solver

Null Geodesics: Hamilton's equations

For a spacetime geometry described by a metric $g_{μν}$ and its inverse $g^{μν}$, null geodesics may be described by the Hamiltonian:

\[H = \frac{1}{2} g^{μν} p_μ p_ν\]

and the associated Hamilton equations:

\[\frac{dx^μ}{dλ} = \frac{∂H}{∂p_μ}\]

\[\frac{dp_μ}{dλ} = - \frac{∂H}{∂x^μ}.\]

The conjugate momenta $p_μ$ are given by:

\[p_μ = g_{μν} \frac{dx^ν}{dλ},\]

and for null geodesics, the initial data satisfies:

\[\left.δ{^i}{_μ} \frac{dx^μ}{dλ}\right|_{λ=0} = {\bf v}^i\]

\[\left. g_{μν}({x}_0) \frac{dx^μ}{dλ}\frac{dx^ν}{dλ} \right|_{λ=0} = 0,\]

where ${\bf v}$ denotes the spatial components of the initial four-velocity vector.

The solution to Hamilton's equations is formally given by $x^μ=x^μ(λ,{x}_0,{\bf v})$ (with $x_0$ denoting the initial position of the geodesic), and since $λ$ (being an affine parameter) can be redefined linearly, it is appropriate to set up the problem so that $λ∈[0,1]$, with $λ=0$ being the initial point and $λ=1$ is the final point.

Implementation

Hamiltonian

The Hamiltonian function takes the form:

squirrel.HamGeo — Function

HamGeo( Z::RealVec , gfunc::Function )

The HamGeo function computes the geodesic Hamiltonian $H = \frac{1}{2} g^{μν} p_μ p_ν$ from the metric $g_{μν}$ provided in the function gfunc and the spacetime position x and momentum p encoded in Z=(x,p).

Hamilton's equations

Hamilton's equations may be written in terms of the phase space coordinate $z^α$, where $z=(x,p)$,

\[\frac{dz^α}{dλ} = J^{αβ} \frac{∂H}{∂z^β}.\]

Here $J^{αβ}$ is the symplectic matrix, which has the block matrix form:

\[J^{\cdot\cdot} = \left[ \begin{array}{cccc} O & I \\ -I & O \end{array} \right],\]

where $O$ is a $4×4$ matrix of zeros and $I$ is the identity matrix. The symplectic matrix is implemented as an operator acting on a vector ${∂H}/{∂z^β}$:

squirrel.Jsympl — Function

Jsympl( Zarg::RealVec )

The Jsympl function effectively applies the symplectic matrix $J^{αβ}$ to the vector Zarg (the vector ${∂H}/{∂z^β}$).

The quantity $J^{αβ} \frac{∂H}{∂z^β}$ is evaluated in the following function:

squirrel.ZdotGeo — Function

ZdotGeo( Z::RealVec , gfunc::Function )

The ZdotGeo function computes the gradient of the Hamiltonian and applies the symplectic operator to the gradient.

Geodesic solver function

Geodesics are solved with the following function, which outputs the endpoint ($λ=1$) of the solution to Hamilton's equations:

squirrel.solveZ — Function

solveZ( Z0::RealVec , gfunc::Function , δ1::Real , δ2::Real ,
        integrator=AutoVern7(Rodas5()) , δt=0 )

The function solveZ performs the integration of geodesics given the initial data Z0, and the metric function gfunc. The integration is performed using the integrator specified by the variable integrator using the relative tolerance parameter δ1 and the absolute tolerance parameter δ2. The parameter δt determines the timestep parameter dt in ODEProblem.

The function solveZ outputs only the endpoint of the solution for the geodesic equation.

Following the recommendations in the ODE Solver documentation for the OrdinaryDiffEq.jl library, the integrators AutoVern7(Rodas5()) and AutoVern9(Rodas5()) are used in squirrel.jl.