Squirrel algorithm

Root finding algorithm

Given some function $f(x)$, the standard approach is the Newton method, which finds the roots according to the prescription:

\[ x_{i+1} = x_i + {\bf J}^{-1}_{i} f(x_i)\]

where ${\bf J}^{-1}_{i}$ is the inverse of the Jacobian matrix ${\bf J}$ for $f(x)$ evaluated at $x_i$. However, in situations which the computation of the Jacobian matrix ${\bf J}$ becomes expensive, one may instead employ the Broyden method, which is a quasi-Newton root finding method. The Broyden algorithm involves first computing the Jacobian matrix ${\bf J}$ and its inverse for $f(x)$. However, instead of computing the Jacobian matrix at each iteration (as is done in the Newton method), the Broyden algorithm updates the (inverse) Jacobian matrix according to the Sherman-Morrison update formula:

\[ {\bf J}^{-1}_{i+1} = {\bf J}^{-1}_{i} + \frac{Δv^{T}_i - {\bf J}^{-1}_{i} Δf_i } {Δv^{T}_i {\bf J}^{-1}_{i} Δf_i } Δv^{T}_i {\bf J}^{-1}_{i},\]

which is implemented in the following function:

squirrel.JiSMU — Function

JiSMU( ΔF::RealVec , Δx::RealVec , Ji::RealMtx )

The function JiSMU implements the Sherman-Morrison update formula.

The Broyden update formula is implemented in the following function:

squirrel.bsolve — Function

bsolve( F::Function , J::RealMtx , f0::RealVec , x0::RealVec ,
        nb::Int=24 )

The function bsolve implements the Broyden algorithm; in particular, it finds the roots of the function F(x), given the Jacobian matrix J and the initial guesses f0 and x0. The parameter nb specifies the maximum number of iterations.

Geodesic endpoints and Jacobian

squirrel.gsolve — Function

gsolve( Xi::RealVec , Vi::RealVec , g::Function , tol::Real , integrator=AutoVern7(Rodas5()) )

The gsolve function takes an initial point Xi and four velocity Vi and computes the endpoint of a future pointing null geodesic in the metric func g. The variable integrator specifies the integration scheme, and tol is the tolerance parameter.

squirrel.zF — Function

zF( Vid::RealVec , Zi::RealMtx , g::Function , tol::Real )

The function zF returns differences between the endpoints of four geodesics for the initial data encoded in Vid and Zi, and the metric function g. The variable tol is the tolerance parameter for the integration. This function vanishes when the four geodesics intersect.

squirrel.gejac — Function

gejac( Xi::RealVec , Vi::RealVec , g::Function , δ::Real )

The function gejac computes the endpoint of a geodesic and its Jacobian. The variables have the same meaning as those in gsolve.

squirrel.geocJ — Function

geocJ( Zi::RealMtx , g::Function , δ::Real )

The function geocJ computes the Jacobian of the function zF from the Jacobian endpoints computed in gejac.

squirrel.idf — Function

idf( Zi::RealMtx , gfunc::Function , tol::Real , nb::Int )

The function idf applies the Broyden algorithm to the function zF, using the Jacobian initially computed with geocJ.

Locator

squirrel.locator4 — Function

locator4( X::RealMtx , Xc::RealVec , gfunc::Function , tol::Real , nb::Int=24 , erc::Bool=false )

The function locator4 computes the intersection point from a set of four emission points X using the guess Xc. The intersection point is computed by first finding the initial data using the function idf, then integrating geodesics to obtain the result.

squirrel.locator — Function

locator( X::RealMtx , gfunc::Function , δ::Real , nb::Int=24 , tpflc::DataType=Double64 , outthresh::Real=1e1 , ne::Int=5 )

The function locator computes the intersection point from a set of ne>4 emission points X by applying locator4 to all combinations of 4 points out of ne in X. A basic outlier detection algorithm (implemented in the function odetc) is applied to reduce errors.