Squirrel algorithm

Description

The squirrel algorithm takes the coordinates of $n_{\rm e} ≥ 4$ emission points, arranged as column vectors in a $4×n_{\rm e}$ matrix $X$, and computes the intersection of future pointing null geodesics defined with respect to a (slightly curved) spacetime metric g.

Null geodesics may be described as solutions to the geodesic equation (implemented here in the form of Hamilton's equations), which may be formally written as the functions:

\[ x^μ_I(λ,{X}_I,{\bf v}_I),\]

where the indices $I∈\{1,2,...,n_{\rm e}\}$ distinguish the emission points $X_I$ and their associated null geodesics, and ${\bf v}_I$ denotes the spatial components of the initial four-velocity vector for the null geodesic (the time component is determined by the requirement that the four-velocity is null).

Given a collection of four geodesic functions $\{x_1,x_2,x_3,x_4\}$ (each having the form $x^μ_I=x^μ_I(λ,{x}_0,{\bf v})$), the condition that they intersect is the vanishing of the following vector-valued function:

\[f := \left( x_1 - x_2 , x_1 - x_3 , x_1 - x_4 \right)\]

which may be formally written as $f=f(X_1,X_2,X_3,X_4,{\bf v}_1,{\bf v}_2,{\bf v}_3,{\bf v}_4)$. A variable counting exercise reveals that $f$ has $12$ components; since there are a total of 12 undetermined quantities in the four initial velocities ${\bf v}_I$ (each of which have three components), the condition $f=0$ can thought of as a set of $12$ equations for the $12$ unknowns $v_I$. The squirrel algorithm seeks to find the roots of the function $f$.

The squirrel algorithm is then summarized:

Apply the flat spacetime algorithm to the emission points $X$ to obtain an initial guess for the zeros of the function $f$.

Apply a root finding algorithm to the function $f$ to obtain the initial velocities $v_I$.

Integrate the geodesics with the resulting initial velocities $v_I$ and emission points $X$ to find the intersection point.

A quasi-Newton Broyden algorithm (which will be described in detail below) is employed to do the root-finding; in such a method, the Jacobian for $f$ is computed once in the first iteration of the root finding algorithm, and is updated in the subsequent iterations. The function $f$ is computed by way of numerical integration of geodesics; if the numerical integration is performed using native Julia libraries, one can compute the Jacobian by way of automatic differentiation.

Geodesic endpoint function

To compute the function $f$, the geodesic endpoint function $x^μ_I=x^μ_I(λ,X_I,{\bf v}_I)|_{λ=1}$ is implemented using the native Julia ODE solvers in the library OrdinaryDiffEq.jl, using the recommended method AutoVern7(Rodas5())

squirrel.gsolve — Function

gsolve( Xi::RealVec , Vi::RealVec , g::Function , tol::Real , integrator=AutoVern7(Rodas5()) )

The gsolve function takes an initial point Xi and four velocity Vi and computes the endpoint of a future pointing null geodesic in the metric func g. The variable integrator specifies the integration scheme, and tol is the tolerance parameter.

With the geodesic endpoint functions $x^μ_I=x^μ_I(λ,X_I,{\bf v}_I)|_{λ=0}$ in hand, one can construct the function $f=f(X_I,{\bf v}_I)$:

squirrel.zF — Function

zF( Vid::RealVec , Zi::RealMtx , g::Function , tol::Real )

The function zF returns differences between the endpoints of four geodesics for the initial data encoded in Vid and Zi, and the metric function g. The variable tol is the tolerance parameter for the integration. This function vanishes when the four geodesics intersect.

Jacobian

Next, one computes the Jacobian of $f$. As mentioned earlier, this is done by way of automatic differentiation, using the library ForwardDiff.jl. Here, the Jacobian matrix of $x^μ_I=x^μ_I(λ,X_I,{\bf v}_I)|_{λ=1}$ (which one may write schematically as $\frac{∂x_I}{∂v_J}$) is computed:

squirrel.gejac — Function

gejac( Xi::RealVec , Vi::RealVec , g::Function , δ::Real )

The function gejac computes the endpoint of a geodesic and its Jacobian. The variables have the same meaning as those in gsolve.

Given $\frac{∂x_I}{∂v_J}$, the Jacobian matrix of the function $f$ may be computed by way of the chain rule, as indicated in the schematic formula:

\[J = \frac{∂f}{∂x_I}\frac{∂x_I}{∂v_J},\]

The following function applies the chain rule to compute the Jacobian matrix:

squirrel.geocJ — Function

geocJ( Zi::RealMtx , g::Function , δ::Real )

The function geocJ computes the Jacobian of the function zF from the Jacobian endpoints computed in gejac.

Root finding algorithm

Given some function $f(x)$, the standard approach is the Newton method, which finds the roots according to the prescription:

\[ x_{i+1} = x_i + {\bf J}^{-1}_{i} f(x_i)\]

where ${\bf J}^{-1}_{i}$ is the inverse of the Jacobian matrix ${\bf J}$ for $f(x)$ evaluated at $x_i$. However, in situations which the computation of the Jacobian matrix ${\bf J}$ becomes expensive, one may instead employ the Broyden method, which is a quasi-Newton root finding method. The Broyden algorithm involves first computing the Jacobian matrix ${\bf J}$ and its inverse for $f(x)$. However, instead of computing the Jacobian matrix at each iteration (as is done in the Newton method), the Broyden algorithm updates the (inverse) Jacobian matrix according to the Sherman-Morrison update formula:

\[ {\bf J}^{-1}_{i+1} = {\bf J}^{-1}_{i} + \frac{Δv^{T}_i - {\bf J}^{-1}_{i} Δf_i } {Δv^{T}_i {\bf J}^{-1}_{i} Δf_i } Δv^{T}_i {\bf J}^{-1}_{i},\]

which is implemented in the following function:

squirrel.JiSMU — Function

JiSMU( ΔF::RealVec , Δx::RealVec , Ji::RealMtx )

The function JiSMU implements the Sherman-Morrison update formula.

The Broyden update formula is implemented in the following function:

squirrel.bsolve — Function

bsolve( F::Function , J::RealMtx , f0::RealVec , x0::RealVec ,
        nb::Int=24 )

The function bsolve implements the Broyden algorithm; in particular, it finds the roots of the function F(x), given the Jacobian matrix J and the initial guesses f0 and x0. The parameter nb specifies the maximum number of iterations.

Initial data finder

The following function makes use of the preceding functions to construct the initial data for the four-velocities:

squirrel.idf — Function

idf( Zi::RealMtx , gfunc::Function , tol::Real , nb::Int )

The function idf applies the Broyden algorithm to the function zF, using the Jacobian initially computed with geocJ.

Locator (Four points)

Finally, the locator4 function computes the intersection point from a set of four emission points X by first employing the idf function to obtain the initial data for the geodesics, and integrates the geodesics (up to unit affine parameter $λ$) to obtain the intersection point:

squirrel.locator4 — Function

locator4( X::RealMtx , Xc::RealVec , gfunc::Function , tol::Real , nb::Int=24 , idv::Bool=false , V::RealMtx=zeros(Float64,4,4) )

The function locator4 computes the intersection point from a set of four emission points X using the guess Xc. The intersection point is computed by first finding the initial data using the function idf, then integrating geodesics to obtain the result. If an improved guess for the four-velocities is available, one can set ìdv=true and specify the four-velocities as column vectors in the matrix V.

Locator ($n_e>4$ emission points)

If more than four emission points are available, the following function can generate $C(n_{\rm e},4)$ sets of emission points (where $C(n,k)$ is the binomial coefficient):

squirrel.combX — Function

combX( X::RealMtx )

The combX function constructs and returns a vector of $4×4$ matrices consisting of all choices of 4 column vectors from the $4×n_{\rm e}$ matrix X.

One feeds the output of combX into the locator4 function to obtain $C(n_{\rm e},4)$ solutions. The following outlier detection formula, which is based on the comparison with median values, can then be used to exclude large errors (the assumption here is that the large errors are infreqent):

squirrel.odetc — Function

odetc( XC::Vector , ξ::Real=1e1 )

The odetc function takes a vector of vectors as input and returns a vector of the vectors that differ by less than a threshold of ξ.

The following function implements the procedure described above, given a matrix of $4×n_{\rm e}$ emission points $X$:

squirrel.locator — Function

locator( X::RealMtx , gfunc::Function , δ::Real , nb::Int=24 , tpflc::DataType=Double64 , outthresh::Real=1e1 , ne::Int=5 )

The function locator computes the intersection point from a set of ne>4 emission points X by applying locator4 to all combinations of 4 points out of ne in X. A basic outlier detection algorithm (implemented in the function odetc) is applied to reduce errors.